requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
django-environ==0.4.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
